Pass maxEvents from user configuration into FFI layer
This also switches to taking Ints everywhere, just as it was before.

+++ glean-core/ffi/src/lib.rs
@@ -208,7 +208,7 @@ pub struct FfiConfiguration<'a> {
     data_dir: FfiStr<'a>,
     package_name: FfiStr<'a>,
     upload_enabled: u8,
-    max_events: Option<&'a i64>,
+    max_events: Option<&'a i32>,
 }
 
 /// Convert the FFI-compatible configuration object into the proper Rust configuration object.
